-New hierarchy: src contains sources, bin contains library file -Compiler: compiler.jar is Google tool -compile.py: python script for compiling
