+++ README.md
@@ -1,4 +1,5 @@
 poyEvo
 ======
+![poyEvo logo](./poyEvo.png?raw=true)
 
 poyEvo is a simple but complete JS library for animation and more
